requirements.txt
flask==2.3.3
azure-ai-translation-text==1.0.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
markdown==3.5.1
lxml==4.9.3
gunicorn==21.2.0
